♻️(candidate-presentation) delete candidate fixtures
