--- src/tycho/tests/candidate/conftest.py
@@ -1,1 +1,0 @@
-from tests.fixtures.candidate_fixtures import *  # noqa: F403
